backend/requirements.txt
# ---- Core API ----
Flask==3.1.2
flask-cors==6.0.2
gunicorn==21.2.0
python-dotenv==1.2.1

# ---- Auth & Security ----
bcrypt==5.0.0
passlib==1.7.4
python-jose==3.5.0
cryptography==46.0.4

# ---- Database ----
pymongo==4.16.0

# ---- HTTP ----
requests==2.32.5
httpx==0.28.1

# ---- Image Processing ----
opencv-python-headless==4.10.0.84
pillow==12.1.0
pi_heif==1.2.0

# ---- AI / ML (PyTorch) ----
torch==2.5.1+cu121
torchvision==0.20.1+cu121
ultralytics==8.4.10
onnx==1.20.1
onnxruntime-gpu==1.23.2

# ---- Utilities ----
numpy==2.2.6
scipy==1.15.3
tqdm==4.67.2
rich==14.3.1
coloredlogs==15.0.1
filelock==3.20.3

# ---- FastAPI & Uvicorn ----
fastapi
uvicorn
passlib[bcrypt]
